requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
distro==1.9.0
dj-database-url==2.1.0
Django==5.2.4
django-admin-honeypot-updated-2021==1.2.0
django-allauth==65.10.0
django-environ==0.12.0
django-ipware==7.0.1
exceptiongroup==1.3.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.98.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
python-ipware==3.0.0
qrcode==8.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.6.0
